+++ sdk/release_notes.md
@@ -3,7 +3,11 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR/#issue)
 -->
-
 ### Microsoft.Azure.Functions.Worker.Sdk 1.16.3 (meta package)
 
 - Update worker.config generation to accurate worker executable name (#1053)
+- Update Microsoft.Azure.Functions.Worker.Sdk.Generators dependency to 1.1.5
+- 
+### Microsoft.Azure.Functions.Worker.Sdk.Generators dependency to 1.1.5
+
+- Adding support for executing functions from referenced assemblies in the optimized executor (#2089)